requirements.txt
python-arango>=7.0.0
requests>=2.28.0
python-dotenv>=0.19.0

